requirements.txt
certifi==2020.12.5
Pillow==8.0.1
pytesseract==0.3.6
fastapi~=0.62.0
uvicorn
setuptools~=51.0.0
requests~=2.25.0
flask~=1.1.2
werkzeug~=1.0.1